+++ src/yadism/structure_functions/nc/f_total.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
 
 
-from . import esf
-from . import esf_result
+from .. import esf
 
 
 class EvaluatedStructureFunctionFtotal(esf.EvaluatedStructureFunction):
@@ -16,9 +15,6 @@ class EvaluatedStructureFunctionFtotal(esf.EvaluatedStructureFunction):
             kinematics : dict
                 the specific kinematic point as a dict with two elements ('x', 'Q2')
     """
-
-    def __init__(self, SF, kinematics: dict):
-        super(EvaluatedStructureFunctionFtotal, self).__init__(SF, kinematics)
 
     def get_result(self):
         """
