+++ src/yadism/structure_functions/nc/fl_heavy.py
@@ -22,7 +22,7 @@ The main reference used is: :cite:`felix-thesis`.
 
 import numpy as np
 
-from .esf import EvaluatedStructureFunctionHeavy as ESFH
+from ..esf import EvaluatedStructureFunctionHeavy as ESFH
 
 
 class EvaluatedStructureFunctionFLheavy(ESFH):
